Address FreeTDS 1.0 compiler warnings.
* Harmonize type usage, via explicit casts where necessary (and safe).
  In particular, ensure that upstream tests' tds_thread_proc routines
  know to return DWORD rather than void* on Windows.
* Ensure that any headers checking (TDS_)HAVE_* #include <config.h> first.
* Keep static assertions warning-free by using _Static_assert where
  available.
* Ensure that members of types that should be unhidden have unhidden
  forward declarations; likewise for their members, recursively.
* Remove (comment out) unused variables and constants.
* Express -2**31 and -2**63 in a manner all compilers accept happily.
* For LLVM Clang, substitute __attribute__((visibility("default")))
  for __attribute__((externally_visible)), which it doesn't support.
* Add no-op default: cases to switch statements that don't explicitly
  cover all of TDS_SERVER_TYPE.
* ctlib.h: #define _CS_CURS_TYPE_* to TDS_CURSOR_STATE_* rather than
  (typelessly) duplicating their values.
* thread.h: Correct tds_thread_proc and TDS_THREAD_PROC_DECLARE; Windows
  threads are expected to return DWORD, not void*.
* dblib/bcp.c (_bcp_get_col_data): Remove a redundant test for
  non-negativity.
* dblib/dblib.c: Express -2?\194?\179?\194?\185 in a manner all compilers accept happily.
* odbc/unittests/attributes.c: Tell the compiler fatal() will never return.
* odbc/unittests/common.h: When the ODBC 3 API is available, favor
  SQLSetStmtAttr over SQLSetStmtOption, which is deprecated.
* tds/convert.c: Rename the string_to_result function to ensure IDEs
  don't conflate it with the identically named wrapper macro.
* tds/gssapi.c: Suppress GSSAPI deprecation warnings on Mac OS X.
* tds/md[45].c: Use TDS_PUT_(U)A4 to populate word32 targets cleanly.
* tds/query.c: #include <sys/socket.h> if available, for send().
* tds/sspi.c: Define _WIN32_WINNT before <config.h>, which might
  otherwise supply a slightly different definition.

TRUNK-0.95: 487511, 554113.  JIRA: CXX-9729.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@80680 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/driver/ftds100/freetds/tds/packet.c
@@ -602,7 +602,7 @@ tds_read_packet(TDSSOCKET * tds)
 	tds->in_len = 0;
 	tds->in_pos = 0;
 	for (p = pkt, end = p+8; p < end;) {
-		int len = tds_connection_read(tds, p, end - p);
+                ssize_t len = tds_connection_read(tds, p, end - p);
 		if (len <= 0) {
 			tds_close_socket(tds);
 			return -1;
@@ -764,7 +764,7 @@ int
 tds_put_cancel(TDSSOCKET * tds)
 {
 	unsigned char out_buf[8];
-	int sent;
+        ssize_t sent;
 
 	out_buf[0] = TDS_CANCEL;	/* out_flag */
 	out_buf[1] = 1;	/* final */
